pyproject.toml
[tool.poetry]
name = "chronos"
version = "0.1.0"
description = ""
authors = ["Shivansh Saini <shivanshs9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
apache-airflow = {extras = ["aws"], version = "^1.10.10"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
